Add tests for updating a service, factor out service tests, update fixture with service to be updated
