+++ fixtures/services.json
@@ -8,5 +8,10 @@
     "name": "test_service_2",
     "url": "localhost:3001/validateCASLogin",
     "adminEmail": "admin@test.com"
+  },
+  {
+    "name": "test_service_3",
+    "url": "localhost:3002/validateCASLogin",
+    "adminEmail": "admin@test.com"
   }
 ]
